refactor(api): Share router implementation
